perf: use async loop to keep track of available space (#840)
* fix: use async loop to keep track of available space

- The changes from 831 work but using a global lock blocks all downloads until the query for free space is ready, even if they will go to a different drive. The internal chunk counter is reset after a single query so, if 2 downloads start at almost the same time, the original issue can still occur.

- We can take advantage of 814 by skipping the querying free space for every file and just query the mount point they are using/will use and reuse the result of the query for every file that uses the same mount point.

This PR adds a helper class that runs a loop every 2 seconds to get the free space of every mount point in the system and keeps a dict with the values. This means we can check the available free space after every chunk written to disk by just querying the internal dict with the mount point of the current file, without blocking.

Loop runs every second and only queries mount points that were used/are being used by some file in the current run.

This uses a new dependency ([psutil](https://github.com/giampaolo/psutil/tree/master)) to query all the available mountpoints in the system

* refactor: move storage manager to its own module

* refactor: take into account running multiple configs

* refactor: make reset aware of its stated

* refactor: log storage status

* refactor: change log formatting

* refactor: add `pause_if_no_free_space`

* perf: simplify stats compute

* refactor: reorder methods

* refactor: add `mount_point` as an attribute  of `MediaItem`

* perf: remove `mountpoint` attribute, use a global cache

* perf: cache `get_mount_point`

* refactor: assert paths are absolute

* refactor: split pause and resume

* fix: concurrent check for pause logic

Do not raise an `InsuficienteFreeSpaceError` if the program was pause lest than 2 seconds ago

* fix: lock ups

Only reset after a config change

* feat: add `pause_on_insufficient_space` config option

* feat: show notification via textual

* refactor: reverse wording of textual-ui cli option

* refactor: sort cli options

* fix: key error while logging

Convert mount to a string before logging

* refactor: flatten storage information before logging

* refactor: show optional msg when paused

* refactor: disable pause on insufficient space

Needs textual UI